+++ src/cli.rs
@@ -1,3 +1,6 @@
+use crate::commands::compile::{
+    cairo_to_sierra::CompileCairoToSierra, sierra_to_casm::CompileSierraToCasm, Compile,
+};
 use clap::{Parser, Subcommand};
 
 #[derive(Parser)]
@@ -10,9 +13,14 @@ pub struct Cli {
 
 #[derive(Debug, Subcommand)]
 pub enum Commands {
-    /// Compile Cairo Smart Contracts
-    #[command()]
-    Compile {},
+    #[clap(about = "Compile Cairo1 contracts.")]
+    Compile(Compile),
+
+    #[clap(about = "Compile Cairo1 contracts to Sierra.")]
+    CompileCairo(CompileCairoToSierra),
+
+    #[clap(about = "Compile Sierra artifacts to Casm.")]
+    CompileSierra(CompileSierraToCasm),
 }
 
 #[test]