+++ src/main.rs
@@ -1,13 +1,28 @@
 mod cli;
 mod commands;
+mod common;
+mod utils;
+
+use anyhow::{Ok, Result};
 
 use clap::Parser;
 use cli::Cli;
+use commands::compile::Compiler;
 
-fn main() {
+fn main() -> Result<()> {
     let cli = Cli::parse();
 
-    match &cli.command {
-        cli::Commands::Compile {} => commands::compile(),
+    match cli.command {
+        cli::Commands::Compile(cmd) => {
+            cmd.run()?;
+        }
+        cli::Commands::CompileCairo(cmd) => {
+            cmd.run()?;
+        }
+        cli::Commands::CompileSierra(cmd) => {
+            cmd.run()?;
+        }
     }
+
+    Ok(())
 }