+++ CHANGELOG.md
@@ -5,3 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Added a command to compile Cairo1 contracts to Sierra (`compile-cairo`)
+- Added a command to compile Sierra artifacts to Casm (`compile-sierra`)
+- Added a command to generate Abi, Sierra, and Casm artifacts from Cairo1 contracts (`compile`)